+++ README.md
@@ -107,3 +107,6 @@ Heroku recomment run rails on Puma server. If you want to try it in development 
 1. Copy `.env.example` file and name it `.env`
 2. `bundle install`
 3. Start rails server with Foreman `foreman start`
+
+## Postgrest start
+`pg_ctl -D /usr/local/var/postgres -l /usr/local/var/postgres/server.log start`
