--- app/serializers/base_serializer.rb
@@ -1,5 +1,0 @@
-class BaseSerializer < Oat::Serializer
-
-  adapter FlatAPI
-
-end
