--- app/serializers/user_serializer.rb
@@ -1,8 +1,0 @@
-class UserSerializer < BaseSerializer
-
-  schema do
-    type 'user'
-
-    map_properties :id, :name, :email, :created_at, :updated_at
-  end
-end
