--- app/serializers/users_serializer.rb
@@ -1,8 +1,0 @@
-class UsersSerializer < BaseSerializer
-
-  schema do
-    type 'users'
-
-    collection :users, item, UserSerializer
-  end
-end
