+++ config/initializers/doorkeeper.rb
@@ -7,7 +7,7 @@ Doorkeeper.configure do
   # This block will be called to check whether the resource owner is authenticated or not.
   resource_owner_authenticator do
     session[:user_return_to] = request.fullpath
-    User.find_by_id(session[:current_user_id]) || redirect_to(public_pages.login_url)
+    Account::User.find_by_id(session[:current_user_id]) || redirect_to(account.login_url)
   end
 
   # If you want to restrict access to the web interface for adding oauth authorized applications, you need to declare the block below.
