+++ engines/account/app/controllers/account/api/v1/credentials_controller.rb
@@ -1,0 +1,24 @@
+module Account
+  module Api
+    module V1
+
+      class CredentialsController < ApplicationController
+        before_action :doorkeeper_authorize!
+
+        # GET /me.json
+        def me
+          unauthorized_request unless current_resource_owner
+          render json: Account::UserSerializer.new(current_resource_owner)
+        end
+
+        private
+
+        # Find the user that owns the access token
+        def current_resource_owner
+          Account::User.find(doorkeeper_token.resource_owner_id) if doorkeeper_token
+        end
+
+      end
+    end
+  end
+end
