+++ engines/account/app/serializers/account/base_serializer.rb
@@ -1,0 +1,5 @@
+module Account
+  class BaseSerializer < Oat::Serializer
+    adapter FlatAPI
+  end
+end
