+++ engines/account/app/serializers/account/users_serializer.rb
@@ -1,0 +1,10 @@
+module Account
+  class UsersSerializer < Account::BaseSerializer
+
+    schema do
+      type 'users'
+
+      collection :users, item, Account::UserSerializer
+    end
+  end
+end
